refactor: catch payment and invoice error at faspi exceptionhandler level
refactor exceptionhandlers into `exception.py` also now always throw
payment error when pay_invoice and invoice errors when create_invoice.

return a status flag with the detailed error message. with a 520
response

ruff
